whisper/python-service-simulator/requirements.txt
numpy==2.2.2
omegaconf==2.3.0
pyzmq==26.2.1
